Start releasing version 3.0.0

+++ src/main/java/rocks/massi/controllers/CacheController.java
@@ -3,8 +3,10 @@ package rocks.massi.controllers;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
+import rocks.massi.authentication.TrollsJwt;
 import rocks.massi.cache.CrawlCache;
 import rocks.massi.data.CacheOperation;
+import rocks.massi.exceptions.AuthenticationException;
 
 import java.io.IOException;
 import java.util.Date;
@@ -19,8 +21,16 @@ public class CacheController {
     @Autowired
     private CrawlCache crawlCache;
 
+    @Autowired
+    private TrollsJwt trollsJwt;
+
+    @CrossOrigin(allowedHeaders = {"Authorization"})
     @DeleteMapping("/purge")
-    public CacheOperation purgeCache() {
+    public CacheOperation purgeCache(@RequestHeader("Authorization") final String authorization) {
+        if (!trollsJwt.checkHeaderWithToken(authorization)) {
+            throw new AuthenticationException("User not authorized");
+        }
+
         try {
             crawlCache.purgeAll();
             return new CacheOperation(true, "", new LinkedList<>());
@@ -30,8 +40,13 @@ public class CacheController {
         }
     }
 
+    @CrossOrigin(allowedHeaders = {"Authorization"})
     @DeleteMapping("/expired")
-    public CacheOperation purgeExpired() {
+    public CacheOperation purgeExpired(@RequestHeader("Authorization") final String authorization) {
+        if (!trollsJwt.checkHeaderWithToken(authorization)) {
+            throw new AuthenticationException("User not authorized");
+        }
+
         try {
             crawlCache.purgeExpired();
             return new CacheOperation(true, "", new LinkedList<>());
@@ -41,8 +56,13 @@ public class CacheController {
         }
     }
 
+    @CrossOrigin(allowedHeaders = {"Authorization"})
     @GetMapping("/get")
-    public CacheOperation getMemoryCache() {
+    public CacheOperation getMemoryCache(@RequestHeader("Authorization") final String authorization) {
+        if (!trollsJwt.checkHeaderWithToken(authorization)) {
+            throw new AuthenticationException("User not authorized");
+        }
+
         List<CacheOperation.CacheEntry> entries = new LinkedList<>();
         crawlCache.getCache().forEach((k, v) -> entries.add(new CacheOperation.CacheEntry(k, v, new Date(v * 1000).toString())));
         return new CacheOperation(true, "", entries);
